wine: updated resources wine-gecko 2.36 & wine-mono 4.5.6
In order that non-stable versions would be able to use the resources,
I moved the resources to a bigger block.

By the way, is it possible to specify a resource with `:optional`?
For wine, some users may want to download the extra packages
manually when needed (wine will pop a window up to ask the user to
install one.)

A link to the sourceforge files list of the wine project is also
added so it would be easier to check for new versions.